refactor: 프로젝트 상세 페이지 내 공통 레이아웃 Section 컴포넌트로 분리(#203)
